my-ai-twin/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
# audioop-lts==0.2.2   # Switched to python 3.11, which has embedded audioop
aws-lambda-powertools==3.19.0
backoff==2.2.1
beautifulsoup4==4.13.5
boto3==1.40.16
boto3-stubs==1.40.16
botocore==1.40.16
botocore-stubs==1.38.46
brotli==1.1.0
bs4==0.0.2
build==1.3.0
cachecontrol==0.14.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cleo==2.1.0
click==8.1.1
cloudpickle==3.1.1
comet-ml==3.51.0
configobj==5.0.9
core-db==2.2.0
core-etl==1.1.3
core-mixins==2.0.3
core-tests==1.1.5
coverage==7.10.3
crashtest==0.4.1
cryptography==45.0.6
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dulwich==0.24.1
emoji==2.14.1
environs==14.3.0
everett==3.1.0
fastapi==0.116.1
fastjsonschema==2.21.2
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
findpython==0.6.3
frozenlist==1.7.0
fsspec==2025.3.0
gdown==5.2.0
google-pasta==0.2.0
#gradio==2.8.0
gradio==5.46.1
gradio-client==1.13.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
groovy==0.1.2
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.8
hpack==4.1.0
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
installer==0.7.0
instructorembedding==1.0.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keyring==25.6.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.147
litellm==1.76.0
lxml==6.0.1
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mock==4.0.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
-e file:///Users/sambapedapalli/work/code/ml-ai-projects/my-ai-twin
mypy-boto3-bedrock-runtime==1.40.14
mypy-extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
openai==1.99.9
opik==1.8.33
orjson==3.11.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.2
pathos==0.3.2
pbs-installer==2025.8.18
pika==1.3.2
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.4
poetry-core==2.1.3
portalocker==2.10.1
pox==0.3.6
ppft==1.7.7
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pygments==2.19.2
pymongo==4.14.1
pypdf==6.0.0
pyproject-hooks==1.2.0
pysocks==1.7.1
pytest==8.4.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.12.1
rapidfuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
ruff==0.12.10
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
sagemaker==2.251.0
sagemaker-core==1.0.54
schema==0.7.7
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.35.0
semantic-version==2.10.0
sentence-transformers==5.1.0
sentry-sdk==2.35.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.43
starlette==0.47.3
structlog==25.4.0
sympy==1.14.0
tblib==3.1.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
trio==0.30.0
trio-websocket==0.12.2
trove-classifiers==2025.8.6.13
typer==0.16.1
types-awscrt==0.27.6
types-s3transfer==0.13.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unstructured==0.18.13
unstructured-client==0.42.3
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
virtualenv==20.32.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
wurlitzer==3.1.1
xattr==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
